Move TimingServiceBase class to FWCore/AbstractServices
