+++ PerfTools/AllocMonitor/plugins/BuildFile.xml
@@ -1,4 +1,5 @@
+<use name="FWCore/AbstractServices"/>
+<use name="FWCore/Concurrency"/>
 <use name="FWCore/MessageLogger"/>
 <use name="FWCore/ServiceRegistry"/>
 <use name="PerfTools/AllocMonitor"/>
-<use name="FWCore/Concurrency"/>
